Refactor CEF implementation modules
