+++ crates/tauri-runtime-cef/src/cef_impl/client/context_menu.rs
@@ -1,0 +1,27 @@
+// Copyright 2019-2024 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+use cef::*;
+
+wrap_context_menu_handler! {
+  pub(super) struct TauriCefContextMenuHandler {
+    devtools_enabled: bool,
+  }
+
+  impl ContextMenuHandler {
+    fn on_before_context_menu(
+      &self,
+      _browser: Option<&mut Browser>,
+      _frame: Option<&mut Frame>,
+      _params: Option<&mut ContextMenuParams>,
+      model: Option<&mut MenuModel>,
+    ) {
+      if !self.devtools_enabled
+        && let Some(model) = model
+      {
+        model.remove_at(model.count() - 1);
+      }
+    }
+  }
+}
