+++ crates/tauri-runtime-cef/src/cef_impl/client/display.rs
@@ -1,0 +1,62 @@
+// Copyright 2019-2024 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+use std::sync::Arc;
+
+use cef::*;
+
+use crate::webview::INITIAL_LOAD_URL;
+
+wrap_display_handler! {
+  pub(super) struct TauriCefDisplayHandler {
+    document_title_changed_handler: Option<Arc<tauri_runtime::webview::DocumentTitleChangedHandler>>,
+    address_changed_handler: Option<Arc<tauri_runtime::webview::AddressChangedHandler>>,
+  }
+
+  impl DisplayHandler {
+    fn on_title_change(
+      &self,
+      _browser: Option<&mut Browser>,
+      title: Option<&CefString>,
+    ) {
+      let Some(handler) = &self.document_title_changed_handler else {
+        return;
+      };
+      let Some(title) = title else {
+        return;
+      };
+
+      handler(title.to_string());
+    }
+
+    fn on_address_change(
+      &self,
+      _browser: Option<&mut Browser>,
+      frame: Option<&mut Frame>,
+      url: Option<&CefString>,
+    ) {
+      // Only fire for main frame URL changes (matches on_before_browse behavior).
+      if let Some(frame) = frame
+        && frame.is_main() == 0
+      {
+        return;
+      }
+      let Some(handler) = &self.address_changed_handler else {
+        return;
+      };
+      let Some(url) = url else {
+        return;
+      };
+      let url = url.to_string();
+
+      if url == INITIAL_LOAD_URL {
+        return;
+      }
+
+      if let Ok(url) = url::Url::parse(&url) {
+        handler(&url);
+      }
+    }
+  }
+}
