+++ crates/tauri-runtime-cef/src/cef_impl/client/load.rs
@@ -1,0 +1,61 @@
+// Copyright 2019-2024 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+use std::sync::Arc;
+
+use cef::*;
+
+wrap_load_handler! {
+  pub(super) struct TauriCefLoadHandler {
+    on_page_load_handler: Option<Arc<tauri_runtime::webview::OnPageLoadHandler>>,
+  }
+
+  impl LoadHandler {
+    fn on_load_start(
+      &self,
+      _browser: Option<&mut Browser>,
+      frame: Option<&mut Frame>,
+      _transition_type: TransitionType,
+    ) {
+      let Some(handler) = &self.on_page_load_handler else {
+        return;
+      };
+      let Some(frame) = frame else {
+        return;
+      };
+
+      if frame.is_main() == 0 {
+        return;
+      }
+
+      let url = cef::CefString::from(&frame.url()).to_string();
+      if let Ok(url) = url::Url::parse(&url) {
+        handler(url, tauri_runtime::webview::PageLoadEvent::Started);
+      }
+    }
+
+    fn on_load_end(
+      &self,
+      _browser: Option<&mut Browser>,
+      frame: Option<&mut Frame>,
+      _http_status_code: ::std::os::raw::c_int,
+    ) {
+      let Some(handler) = &self.on_page_load_handler else {
+        return;
+      };
+      let Some(frame) = frame else {
+        return;
+      };
+
+      if frame.is_main() == 0 {
+        return;
+      }
+
+      let url = cef::CefString::from(&frame.url()).to_string();
+      if let Ok(url) = url::Url::parse(&url) {
+        handler(url, tauri_runtime::webview::PageLoadEvent::Finished);
+      }
+    }
+  }
+}
