+++ crates/tauri-runtime-cef/src/cef_impl/client/process.rs
@@ -1,0 +1,66 @@
+// Copyright 2019-2024 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+use std::sync::{
+  Arc,
+  atomic::{AtomicBool, Ordering},
+};
+#[cfg(not(target_os = "macos"))]
+use std::time::Duration;
+
+use cef::*;
+use tauri_runtime::UserEvent;
+
+use crate::runtime::{Message, RuntimeContext};
+
+wrap_browser_process_handler! {
+  pub(crate) struct TauriCefBrowserProcessHandler<T: UserEvent> {
+    context: RuntimeContext<T>,
+    context_initialized: Arc<AtomicBool>,
+    deep_link_schemes: Vec<String>,
+  }
+
+  impl BrowserProcessHandler {
+    fn on_context_initialized(&self) {
+      self.context_initialized.store(true, Ordering::SeqCst);
+      self.context.proxy.wake_up();
+    }
+
+    fn on_schedule_message_pump_work(&self, delay_ms: i64) {
+      #[cfg(target_os = "macos")]
+      {
+        self.context.cef_pump.schedule_message_pump_work(delay_ms);
+      }
+      #[cfg(not(target_os = "macos"))]
+      {
+        let delay = Duration::from_millis(delay_ms.max(0) as u64);
+        let _ = self.context.sender.send(Message::CefWork(delay));
+        self.context.proxy.wake_up();
+      }
+    }
+
+    fn on_already_running_app_relaunch(
+      &self,
+      command_line: Option<&mut CommandLine>,
+      _current_directory: Option<&CefString>,
+    ) -> std::os::raw::c_int {
+      let Some(command_line) = command_line else {
+        return 0;
+      };
+      let mut list = CefStringList::new();
+      command_line.arguments(Some(&mut list));
+      let args: Vec<String> = list.into_iter().collect();
+      if let Ok(url) = url::Url::parse(&args[0]) {
+        let scheme = url.scheme().to_string();
+        if self.deep_link_schemes.iter().any(|s| s == &scheme) {
+          let _ = self.context.sender.send(Message::Opened(vec![url]));
+          self.context.proxy.wake_up();
+          return 1;
+        }
+      }
+      // TODO: add event
+      1
+    }
+  }
+}
